+++ pyraxshell/__main__.py
@@ -25,6 +25,7 @@ from globals import CONFIG_FILE
 from utility import check_dir_home
 from sessions import Sessions
 from db import DB
+import log
 
 def main():
     # check '~/.pyraxshell' and config files  exist, if not then create it
@@ -35,11 +36,15 @@ def main():
         DB()
         Sessions.Instance().create_table_sessions()  # @UndefinedVariable
         Sessions.Instance().create_table_commands()  # @UndefinedVariable
+        # create default logging configuration file 
+        log.Log()
+        # create default configuration file
+        Configuration.Instance()  # @UndefinedVariable
         sys.exit(0)
     
     # ########################################
     # LOGGING
-    utility.logging_start()  # @UndefinedVariable
+    log.Log().start()
     logging.debug('starting')
     
     # ########################################
@@ -59,9 +64,9 @@ def main():
     # ########################################
     # DO STUFF
     # handle configuration
-    if cfg.pyrax_http_debug:
+    if cfg.pyrax_http_debug == True:
         pyrax.set_http_debug(True)
-    if cfg.pyrax_no_verify_ssl:
+    if cfg.pyrax_no_verify_ssl == True:
         # see: https://github.com/rackspace/pyrax/issues/187
         pyrax.set_setting("verify_ssl", False)
     # main loop