+++ RELEASENOTES.md
@@ -1,9 +1,16 @@
 Release notes for pyraxshell
 =======
 
-## 2013.09.23 - Version 0.2.3
+## 2013.XX.XX - Version 0.2.3
   * Fixed issue #11
   * Fixed issue #9
+  * Added SQLite3 support
+  * Added '~/.pyraxshell' directory and config files support
+  * Fixed issue #14
+  * Added sessions support
+  * Added 'r' (record) function to log messages to file and sessions to db
+  * Fixed non-interactive command output
+  * Implemented main and loggin configuration files support
 
 ## 2013.09.22 - Version 0.2.2
   * Fixed issue #3